frontend/requirements.txt
Flask
google-cloud-pubsub
gunicorn
